# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "DentApp API",
        "description": "API aplikacji DentApp",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/message-group": {
            "get": {
                "tags": [
                    "MessageGroup"
                ],
                "summary": "Pobiera grupy wiadomości zalogowanego użytkownika",
                "operationId": "46816d3660e66c06455f646596383a34",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/PaginatedResponse"
                                        },
                                        {
                                            "properties": {
                                                "data": {
                                                    "type": "array",
                                                    "items": {
                                                        "$ref": "#/components/schemas/MessageGroupResource"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    ]
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "MessageGroup"
                ],
                "summary": "Tworzy nową grupę wiadomości",
                "operationId": "7660c6d988b2eab512b461f32cb6de30",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "user_uuids"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Zespół"
                                    },
                                    "user_uuids": {
                                        "description": "Min. jeden użytkownik (inny niż twórca)",
                                        "type": "array",
                                        "items": {
                                            "type": "string",
                                            "format": "uuid"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Utworzono",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/MessageGroupResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Błąd walidacji"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/message-group/{uuid}": {
            "get": {
                "tags": [
                    "MessageGroup"
                ],
                "summary": "Pobiera szczegóły grupy wiadomości wraz z członkami",
                "operationId": "a3394ec2ea47284e2513012835d28d9a",
                "parameters": [
                    {
                        "name": "uuid",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/MessageGroupResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Nie znaleziono"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "MessageGroup"
                ],
                "summary": "Zmienia nazwę grupy wiadomości",
                "operationId": "01fe8881ff7bd09ae592672680fbc62b",
                "parameters": [
                    {
                        "name": "uuid",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/MessageGroupResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Nie można edytować grupy domyślnej"
                    },
                    "404": {
                        "description": "Nie znaleziono"
                    },
                    "422": {
                        "description": "Błąd walidacji"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "MessageGroup"
                ],
                "summary": "Usuwa grupę wiadomości",
                "operationId": "d929f9feabc7887af0aab0369127493e",
                "parameters": [
                    {
                        "name": "uuid",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Usunięto"
                    },
                    "403": {
                        "description": "Nie można usunąć grupy domyślnej"
                    },
                    "404": {
                        "description": "Nie znaleziono"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/message-group/{uuid}/users/{userUuid}": {
            "post": {
                "tags": [
                    "MessageGroup"
                ],
                "summary": "Dodaje użytkownika do grupy",
                "operationId": "9487f5f9c39f44dfab52633df0230127",
                "parameters": [
                    {
                        "name": "uuid",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    },
                    {
                        "name": "userUuid",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/MessageGroupResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Nie można edytować grupy domyślnej"
                    },
                    "404": {
                        "description": "Nie znaleziono"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "MessageGroup"
                ],
                "summary": "Usuwa użytkownika z grupy (min. 2 członków musi pozostać)",
                "operationId": "356f89b36a541153c2d8dfd8f49321ca",
                "parameters": [
                    {
                        "name": "uuid",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    },
                    {
                        "name": "userUuid",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Usunięto"
                    },
                    "403": {
                        "description": "Nie można edytować grupy domyślnej lub grupa miałaby tylko 1 członka"
                    },
                    "404": {
                        "description": "Nie znaleziono"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/message-group/{uuid}/messages": {
            "get": {
                "tags": [
                    "MessageGroup"
                ],
                "summary": "Pobiera wiadomości należące do grupy wraz z załączonymi plikami",
                "operationId": "6828cf755200692934d070834baf8c3c",
                "parameters": [
                    {
                        "name": "uuid",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/PaginatedResponse"
                                        },
                                        {
                                            "properties": {
                                                "data": {
                                                    "type": "array",
                                                    "items": {
                                                        "$ref": "#/components/schemas/MessageResource"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Nie znaleziono"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "PaginatedResponse": {
                "properties": {
                    "current_page": {
                        "type": "integer"
                    },
                    "last_page": {
                        "type": "integer"
                    },
                    "per_page": {
                        "type": "integer"
                    },
                    "total": {
                        "type": "integer"
                    },
                    "from": {
                        "type": "integer"
                    },
                    "to": {
                        "type": "integer"
                    }
                },
                "type": "object"
            },
...
            "UserResource": {
                "properties": {
                    "uuid": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "first_name": {
                        "type": "string"
                    },
                    "last_name": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "nullable": true
                    },
                    "private_email": {
                        "type": "string",
                        "format": "email"
                    },
                    "is_active": {
                        "type": "boolean"
                    },
                    "pwz_numer": {
                        "type": "string",
                        "nullable": true
                    },
                    "avatar_path": {
                        "type": "string",
                        "nullable": true
                    },
                    "background_path": {
                        "type": "string",
                        "nullable": true
                    },
                    "private_phone_number": {
                        "type": "string",
                        "nullable": true
                    },
                    "phone_number": {
                        "type": "string",
                        "nullable": true
                    },
                    "job_positions": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/JobPositionResource"
                        }
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
            "JobPositionResource": {
                "properties": {
                    "uuid": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "name": {
                        "type": "string"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
            "MessageResource": {
                "properties": {
                    "uuid": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "user_uuid": {
                        "type": "string",
                        "format": "uuid",
                        "nullable": true
                    },
                    "recipient_user_uuid": {
                        "type": "string",
                        "format": "uuid",
                        "nullable": true
                    },
                    "message_group_uuid": {
                        "type": "string",
                        "format": "uuid",
                        "nullable": true
                    },
                    "message": {
                        "type": "string"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
            "MessageGroupResource": {
                "properties": {
                    "uuid": {
                        "description": "Summary of MessageGroupResource",
                        "type": "string",
                        "format": "uuid"
                    },
                    "name": {
                        "type": "string"
                    },
                    "is_default": {
                        "type": "boolean"
                    },
                    "creator": {
                        "oneOf": [
                            {
                                "$ref": "#/components/schemas/UserResource"
                            }
                        ],
                        "nullable": true
                    },
                    "users": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/UserResource"
                        }
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Token Sanctum — przekaż w nagłówku: Authorization: Bearer {token}",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}